Added interface for the hash generator
